requirements.txt
asgiref==3.7.2
astroid==2.15.6
black==23.9.1
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
dill==0.3.7
distlib==0.3.7
Django==4.2.5
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.3
flake8==6.1.0
identify==2.5.27
isort==5.12.0
lazy-object-proxy==1.4.3
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.2
phonenumbers==8.13.20
platformdirs==3.10.0
pre-commit==3.4.0
psycopg2==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.24.5
wrapt==1.12.1
